requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
aiofiles=22.1.0=py311hca03da5_0
aiosqlite=0.18.0=py311hca03da5_0
anyio=3.5.0=py311hca03da5_0
appnope=0.1.2=py311hca03da5_1001
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py311h80987f9_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=23.1.0=py311hca03da5_0
babel=2.11.0=py311hca03da5_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.12.2=py311hca03da5_0
blas=1.0=openblas
bleach=4.1.0=pyhd3eb1b0_0
bottleneck=1.3.5=py311ha0d4635_0
brotli=1.0.9=h1a28f6b_7
brotli-bin=1.0.9=h1a28f6b_7
brotli-python=1.0.9=py311h313beb8_7
bzip2=1.0.8=h620ffc9_4
ca-certificates=2023.08.22=hca03da5_0
certifi=2023.11.17=py311hca03da5_0
cffi=1.16.0=py311h80987f9_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
comm=0.1.2=py311hca03da5_0
contourpy=1.0.5=py311h48ca7d4_0
cryptography=41.0.3=py311hd4332d6_0
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h9131b1a_1
debugpy=1.6.7=py311h313beb8_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py311hca03da5_0
et_xmlfile=1.1.0=py311hca03da5_0
executing=0.8.3=pyhd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h1192e45_0
gettext=0.21.0=h13f89a0_1
giflib=5.2.1=h80987f9_3
glib=2.69.1=h514c7bf_2
gst-plugins-base=1.14.1=h313beb8_1
gstreamer=1.14.1=h80987f9_1
gurobi=10.0.3=py311_0
icu=73.1=h313beb8_0
idna=3.4=py311hca03da5_0
ipykernel=6.25.0=py311hb6e6a13_0
ipython=8.15.0=py311hca03da5_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=8.0.4=py311hca03da5_0
jedi=0.18.1=py311hca03da5_1
jinja2=3.1.2=py311hca03da5_0
jpeg=9e=h80987f9_1
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.19.2=py311hca03da5_0
jsonschema-specifications=2023.7.1=py311hca03da5_0
jupyter=1.0.0=py311hca03da5_8
jupyter_client=7.4.9=py311hca03da5_0
jupyter_console=6.6.3=py311hca03da5_0
jupyter_core=5.5.0=py311hca03da5_0
jupyter_events=0.8.0=py311hca03da5_0
jupyter_server=1.23.4=py311hca03da5_0
jupyter_server_fileid=0.9.0=py311hca03da5_0
jupyter_server_ydoc=0.8.0=py311hca03da5_1
jupyter_ydoc=0.2.4=py311hca03da5_0
jupyterlab=3.6.3=py311hca03da5_0
jupyterlab_pygments=0.2.2=py311hca03da5_0
jupyterlab_server=2.25.1=py311hca03da5_0
jupyterlab_widgets=3.0.9=py311hca03da5_0
kiwisolver=1.4.4=py311h313beb8_0
krb5=1.20.1=hf3e1bf2_1
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libbrotlicommon=1.0.9=h1a28f6b_7
libbrotlidec=1.0.9=h1a28f6b_7
libbrotlienc=1.0.9=h1a28f6b_7
libclang=14.0.6=default_h1b80db6_1
libclang13=14.0.6=default_h24352ff_1
libcxx=14.0.6=h848a8c0_0
libdeflate=1.17=h80987f9_1
libedit=3.1.20221030=h80987f9_0
libffi=3.4.4=hca03da5_0
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libiconv=1.16=h1a28f6b_2
libllvm14=14.0.6=h7ec7a93_3
libopenblas=0.3.21=h269037a_0
libpng=1.6.39=h80987f9_0
libpq=12.15=h02f6b3c_1
libsodium=1.0.18=h1a28f6b_0
libtiff=4.5.1=h313beb8_0
libwebp=1.3.2=ha3663a8_0
libwebp-base=1.3.2=h80987f9_0
libxml2=2.10.4=h0dcf63f_1
llvm-openmp=14.0.6=hc6e5704_0
lz4-c=1.9.4=h313beb8_0
markupsafe=2.1.1=py311h80987f9_0
matplotlib=3.8.0=py311hca03da5_0
matplotlib-base=3.8.0=py311h7aedaa7_0
matplotlib-inline=0.1.6=py311hca03da5_0
mistune=2.0.4=py311hca03da5_0
munkres=1.1.4=py_0
mysql=5.7.24=ha71a6ea_2
nbclassic=1.0.0=py311hca03da5_0
nbclient=0.8.0=py311hca03da5_0
nbconvert=7.10.0=py311hca03da5_0
nbformat=5.9.2=py311hca03da5_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.5.6=py311hca03da5_0
notebook=6.5.4=py311hca03da5_1
notebook-shim=0.2.3=py311hca03da5_0
numexpr=2.8.7=py311h6dc990b_0
numpy=1.26.0=py311he598dae_0
numpy-base=1.26.0=py311hfbfe69c_0
openjpeg=2.3.0=h7a6adac_2
openpyxl=3.0.10=py311h80987f9_0
openssl=3.0.12=h1a28f6b_0
packaging=23.1=py311hca03da5_0
pandas=2.1.1=py311h7aedaa7_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=hc377ac9_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=10.0.1=py311h3b245a6_0
pip=23.3=py311hca03da5_0
platformdirs=3.10.0=py311hca03da5_0
ply=3.11=py311hca03da5_0
prometheus_client=0.14.1=py311hca03da5_0
prompt-toolkit=3.0.36=py311hca03da5_0
prompt_toolkit=3.0.36=hd3eb1b0_0
psutil=5.9.0=py311h80987f9_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.15.1=py311hca03da5_1
pyopenssl=23.2.0=py311hca03da5_0
pyparsing=3.0.9=py311hca03da5_0
pyqt=5.15.10=py311h313beb8_0
pyqt5-sip=12.13.0=py311h80987f9_0
pysocks=1.7.1=py311hca03da5_0
python=3.11.5=hb885b13_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py311hca03da5_0
python-json-logger=2.0.7=py311hca03da5_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytz=2023.3.post1=py311hca03da5_0
pyyaml=6.0.1=py311h80987f9_0
pyzmq=23.2.0=py311h313beb8_0
qt-main=5.15.2=h0917680_10
qtconsole=5.5.0=py311hca03da5_0
qtpy=2.4.1=py311hca03da5_0
readline=8.2=h1a28f6b_0
referencing=0.30.2=py311hca03da5_0
requests=2.31.0=py311hca03da5_0
rfc3339-validator=0.1.4=py311hca03da5_0
rfc3986-validator=0.1.1=py311hca03da5_0
rpds-py=0.10.6=py311hf0e4da2_0
send2trash=1.8.2=py311hca03da5_0
setuptools=68.0.0=py311hca03da5_0
sip=6.7.12=py311h313beb8_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.2.0=py311hca03da5_1
soupsieve=2.5=py311hca03da5_0
sqlite=3.41.2=h80987f9_0
stack_data=0.2.0=pyhd3eb1b0_0
terminado=0.17.1=py311hca03da5_0
tinycss2=1.2.1=py311hca03da5_0
tk=8.6.12=hb8d0fd4_0
tornado=6.3.3=py311h80987f9_0
traitlets=5.7.1=py311hca03da5_0
typing-extensions=4.7.1=py311hca03da5_0
typing_extensions=4.7.1=py311hca03da5_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.18=py311hca03da5_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py311hca03da5_1
websocket-client=0.58.0=py311hca03da5_4
wheel=0.41.2=py311hca03da5_0
widgetsnbextension=4.0.5=py311hca03da5_0
xz=5.4.2=h80987f9_0
y-py=0.5.9=py311ha6e5c4f_0
yaml=0.2.5=h1a28f6b_0
ypy-websocket=0.8.2=py311hca03da5_0
zeromq=4.3.4=hc377ac9_0
zlib=1.2.13=h5a0b063_0
zstd=1.5.5=hd90d995_0
